+++ datanode/networkhistory/store/store.go
@@ -102,26 +102,30 @@ type Store struct {
 // issue when running tests as we only have one IPFS node instance when running datanode.
 var plugins *loader.PluginLoader
 
-func New(ctx context.Context, log *logging.Logger, chainID string, cfg Config, networkHistoryHome string,
-	wipeOnStartup bool, maxMemoryPercent uint8,
+func New(ctx context.Context, log *logging.Logger, chainID string, cfg Config, networkHistoryHome string, maxMemoryPercent uint8,
 ) (*Store, error) {
 	if log.IsDebug() {
 		ipfslogging.SetDebugLogging()
 	}
 
-	networkHistoryStorePath := filepath.Join(networkHistoryHome, "store")
+	storePath := filepath.Join(networkHistoryHome, "store")
 
 	p := &Store{
 		log:        log,
 		cfg:        cfg,
-		indexPath:  filepath.Join(networkHistoryStorePath, "index"),
-		stagingDir: filepath.Join(networkHistoryStorePath, "staging"),
-		ipfsPath:   filepath.Join(networkHistoryStorePath, "ipfs"),
+		indexPath:  filepath.Join(storePath, "index"),
+		stagingDir: filepath.Join(storePath, "staging"),
+		ipfsPath:   filepath.Join(storePath, "ipfs"),
 	}
 
-	err := p.setupPaths(networkHistoryStorePath, wipeOnStartup)
+	err := os.MkdirAll(p.indexPath, os.ModePerm)
 	if err != nil {
-		return nil, fmt.Errorf("failed to setup paths:%w", err)
+		return nil, fmt.Errorf("failed to create index path:%w", err)
+	}
+
+	err = os.MkdirAll(p.stagingDir, os.ModePerm)
+	if err != nil {
+		return nil, fmt.Errorf("failed to create staging path:%w", err)
 	}
 
 	idxLog := log.With(logging.String("component", "index"))
@@ -480,27 +484,6 @@ func (p *Store) rewriteZipWithMetadata(oldZip string, metadata segment.MetaData)
 	}
 
 	return tmpfile.Name(), nil
-}
-
-func (p *Store) setupPaths(networkHistoryStorePath string, wipeOnStartup bool) error {
-	if wipeOnStartup {
-		err := os.RemoveAll(networkHistoryStorePath)
-		if err != nil {
-			return fmt.Errorf("failed to remove dir:%w", err)
-		}
-	}
-
-	err := os.MkdirAll(p.indexPath, os.ModePerm)
-	if err != nil {
-		return fmt.Errorf("failed to create index path:%w", err)
-	}
-
-	err = os.MkdirAll(p.stagingDir, os.ModePerm)
-	if err != nil {
-		return fmt.Errorf("failed to create staging path:%w", err)
-	}
-
-	return nil
 }
 
 func (p *Store) GetSegmentForHeight(toHeight int64) (segment.Full, error) {